Expression Analysis

StructuredQueries.get_res_field — Function.

Extract the assigned column's name from an assignment-like expression.

Arguments:

e::Expr: An assignment-like expression, which is either a top-level assignment expression, which might look like col3 = f(col1) + g(col2); or a function/macro's keyword argument expression, which might look like foo(col3 = f(col1) + g(col2)).

Returns:

s::Symbol: A symbol specifying the column name that will be assigned to.

source

StructuredQueries.get_value_expr — Function.

Extract the value-defining sub-expression from an assignment-like expression.

Arguments:

e_in::Expr: An assignment-like expression, which is either a top-level expression like col3 = f(col1) + g(col2) or a function/macro's keyword argument like foo(col3 = f(col1) + g(col2)).

Returns:

e_out::Any: A value-defining expression that will be used to compute the value assigned to the column implied by. May be a literal, a raw symbol or a full Expr object.

source

StructuredQueries.find_symbols! — Function.

Recursively descends an expression's AST to find all of the symbols contained in it. Inserts any unquoted symbols that are found into the set argument, s. Note that this function is not designed to handle assignment-like expressions: it is intended for application to value expressions only.

Arguments:

s::Set{Symbol}: A set of symbols that will be mutated whenever any new symbols are found.

e::Any: An expression-like object that will be descended through to find new symbols.

Returns:

Void: This function is used exclusively to mutate the argument s.

source

StructuredQueries.find_symbols — Function.

Recursively descends an expression's AST to find all of the symbols contained in it. Returns all found symbols in a Set{Symbol} object.

Arguments:

e::Any: An expression-like object that will be descended through to find new symbols.

Returns:

s::Set{Symbol}: A set containing all of the symbols found by descending through the expression-like object's AST.

source

StructuredQueries.map_symbols — Function.

Produce a mapping from symbols to numeric indices and a reverse mapping from numeric indices to symbols.

Arguments:

s::Set{Symbol}: A set of symbols that should be assigned numeric indices.

Returns:

mapping::Dict{Symbol, Int}: A mapping from symbols to indices.

reverse_mapping::Vector{Symbol}: A mapping from indices to symbols.

source

StructuredQueries.replace_symbols — Function.

Traverse an AST-like object and replace a fixed set of symbols with tuple-indexing expressions.

Arguments:

e::Any: An AST-like object.

mapping:Dict{Symbol, Int}: A mapping from symbols to numeric indices.

tuple_name::Symbol: The name of the tuple that will be indexed into.

Returns:

new_e::Any: A new AST-like object with all symbols replaced with tuple-indexing operations.

source

StructuredQueries.build_kernel_ex! — Function.

build_kernel_ex!(e, paramters)

Return an Expr to define a (tuple-argument) lambda whose body reflects the structure of e. Also push any query parameters found while traversing e to parameters.

source