Graph Generation

StructuredQueries.QUERYNODE — Constant.

QUERYNODE

Internal map from manipulation verb names (as Symbols) to tuples (T<:QueryNode, H<:QueryHelper).

source

StructuredQueries.gen_graph_ex — Function.

gen_graph_ex(qry)::Expr

Return an Expr to produce a graph representation of qry.

source

StructuredQueries.gen_helpers_ex — Function.

gen_helpers_ex(H, args)::Expr

Return an Expr to produce a Vector{H<:QueryHelper}, for which each element is derived from a query argument in args.

source

StructuredQueries.gen_helper_ex — Function.

gen_helper_ex(H, arg)::Expr

Return an Expr to produce an H <: QueryHelper object from the query argument arg.

source