QueryHelper

StructuredQueries.QueryHelper — Type.

abstract QueryHelper

Leaf subtypes T <: QueryHelper contain data extracted from query arguments and stored in fields of T as resources for collection machineries.

source

StructuredQueries.parts — Function.

parts(helper::QueryHelper)

Return the parts of the helper (this is usually a tuple of useful things, e.g. a kernel generated from a query argument).

source