QueryNode

StructuredQueries.DataNode — Type.

DataNode <: QueryNode

Represent a data source in a QueryNode manipulation graph.

Note: the "leaves" of any QueryNode graph must be DataNodes.

source

Base.isequal — Method.

Base.isequal(dn1::DataNode, dn2::DataNode)::Bool

Test two DataNodes for equality. This "loose" form of DataNode equality is satisfied when both dn1 and dn2 are empty, as in the case of a base DataNode of a Query graph that has not yet been collected against.

source

Base.isequal — Method.

Base.isequal{T<:QueryNode}(q1::T, q2::T)::Bool

Test two QueryNode objects for equality.

Note that the result depends only on the input and args fields of each q1 and q2; the contents of the helpers and parameters fields are not compared. This "loose" form of QueryNode equality is therefore only determined by the content of the expression passed to @query and reflects the expectation that the same query (as passed to @query) twice should produce Query objects that satisfy isequal.

source

Base.isequal — Method.

Base.isequal{T<:JoinNode}(q1::T, q2::T)::Bool

Test two JoinNode objects for equality.

This result depends only on the input1, input2 and argsfields of each q1 and q2; the contents of the helpers and parameters fields are not compared.

source