Public Documentation

This page contains documentation for StructuredQueries.jl's public interface.

See the Internal Documentation for documentation of StructuredQueries.jl's internals.

Contents

Public Documentation

Contents

Index

Public Interface

Index

StructuredQueries.Query

Base.collect

StructuredQueries.graph

StructuredQueries.source

StructuredQueries.@collect

StructuredQueries.@query

Public Interface

StructuredQueries.@query — Macro.

@query(qry)

Return a Query object that represents the query structure of qry.

source

StructuredQueries.@collect — Macro.

@collect(qry)

Like @query, but automatically collects the resulting Query object.

source

StructuredQueries.Query — Type.

Query

Wrap a QueryNode graph that represents the structure of a query passed to the @query macro.

source

Base.collect — Method.

Base.collect(q::Query)

Collect a qry against the source wrapped in the base DataNode of qry.graph.

source

StructuredQueries.source — Function.

source(q::Query)

Return the data source(s) against which q is to be collected.

source

StructuredQueries.graph — Function.

graph(q::Query)

Return the QueryNode graph representation of the query that produced q.

source